replaced 'int64_t more' with 'int more' in C/C++ Multipart examples
Using int64_t on architectures where 'int' is 32bit yields unknown results

+++ articles/src/multithreading/zhelpers.h
@@ -119,7 +119,7 @@ s_dump (void *socket)
         }
         printf ("\n");
 
-        int64_t more;           //  Multipart detection
+        int more;        //  Multipart detection
         size_t more_size = sizeof (more);
         zmq_getsockopt (socket, ZMQ_RCVMORE, &more, &more_size);
         zmq_msg_close (&message);
